requirements.txt
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-JWT==0.3.2
Flask-Login==0.4.0
Flask-SQLAlchemy==2.3.1
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyJWT==1.4.2
PyYAML==3.12
requests==2.18.4
schema==0.6.6
simplejson==3.11.1
slacker==0.9.60
SQLAlchemy==1.1.14
urllib3==1.22
Werkzeug==0.12.2
python-dateutil==2.6.1
